feat: change all mentions of Preview to Skylark Preview, add Popup component, fix detection of unauthenticated user
