+++ src/content.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-import { PreviewStatus, PreviewStatusLine } from "./components/previewStatus";
+import { PreviewStatus } from "./components/previewStatus";
 
 const root = document.createElement("div");
 root.id = "skylark-preview-extension";
@@ -9,7 +9,6 @@ document.body.appendChild(root);
 
 ReactDOM.createRoot(root).render(
   <React.StrictMode>
-    <PreviewStatusLine />
     <PreviewStatus />
   </React.StrictMode>,
 );
